+++ src/datumaro/experimental/categories.py
