+++ src/datumaro/experimental/legacy/annotation_converters.py
@@ -11,8 +11,7 @@ from datumaro import Annotation, AnnotationType, Bbox, CategoriesInfo, DatasetIt
 from datumaro import Dataset as LegacyDataset
 from datumaro import LabelCategories as LegacyLabelCategories
 from datumaro.components.annotation import ExtractedMask, RotatedBbox
-from datumaro.util.mask_tools import generate_colormap
-from datumaro.v2 import (
+from datumaro.experimental import (
     AttributeInfo,
     BBoxField,
     Dataset,
@@ -27,8 +26,14 @@ from datumaro.v2 import (
     polygon_field,
     rotated_bbox_field,
 )
-from datumaro.v2.categories import LabelCategories, MaskCategories, RgbColor
-from datumaro.v2.fields import EllipseField, instance_mask_callable_field, keypoints_field, mask_callable_field
+from datumaro.experimental.categories import LabelCategories, MaskCategories, RgbColor
+from datumaro.experimental.fields import (
+    EllipseField,
+    instance_mask_callable_field,
+    keypoints_field,
+    mask_callable_field,
+)
+from datumaro.util.mask_tools import generate_colormap
 
 
 class ForwardAnnotationConverter(ABC):
@@ -143,7 +148,7 @@ class ForwardBboxAnnotationConverter(ForwardAnnotationConverter):
             attributes[self.name_prefix + "labels"] = self.bbox_labels_attribute
         return attributes
 
-    def convert_annotations(self, annotations: list[Annotation], item: DatasetItem) -> dict[str, Any]:
+    def convert_annotations(self, annotations: list[Annotation], item: DatasetItem) -> dict[str, Any]:  # noqa: ARG002
         bboxes: list[list[float]] = []
         labels: list[int | None] = []
 
@@ -226,7 +231,7 @@ class ForwardRotatedBboxAnnotationConverter(ForwardAnnotationConverter):
             attributes[self.name_prefix + "labels"] = self.rotated_bbox_labels_attribute
         return attributes
 
-    def convert_annotations(self, annotations: list[Annotation], item: DatasetItem) -> dict[str, Any]:
+    def convert_annotations(self, annotations: list[Annotation], item: DatasetItem) -> dict[str, Any]:  # noqa: ARG002
         rotated_bboxes: list[list[float]] = []
         labels: list[int | None] = []
 
@@ -309,7 +314,7 @@ class ForwardPolygonAnnotationConverter(ForwardAnnotationConverter):
             attributes[self.name_prefix + "labels"] = self.polygon_labels_attribute
         return attributes
 
-    def convert_annotations(self, annotations: list[Annotation], item: DatasetItem) -> dict[str, Any]:
+    def convert_annotations(self, annotations: list[Annotation], item: DatasetItem) -> dict[str, Any]:  # noqa: ARG002
         polygons: list[list[float]] = []
         labels: list[int | None] = []
 
@@ -382,7 +387,7 @@ class ForwardLabelAnnotationConverter(ForwardAnnotationConverter):
     def get_schema_attributes(self) -> dict[str, AttributeInfo]:
         return {self.name_prefix + "label": self.label_attribute}
 
-    def convert_annotations(self, annotations: list[Annotation], item: DatasetItem) -> dict[str, Any]:
+    def convert_annotations(self, annotations: list[Annotation], item: DatasetItem) -> dict[str, Any]:  # noqa: ARG002
         labels = [ann for ann in annotations if isinstance(ann, Label)]
         result = {}
         if len(labels) > 0:
@@ -452,7 +457,7 @@ class ForwardKeypointAnnotationConverter(ForwardAnnotationConverter):
             attributes[self.name_prefix + "labels"] = self.keypoints_labels_attribute
         return attributes
 
-    def convert_annotations(self, annotations: list[Annotation], item: DatasetItem) -> dict[str, Any]:
+    def convert_annotations(self, annotations: list[Annotation], item: DatasetItem) -> dict[str, Any]:  # noqa: ARG002
         keypoints = [ann for ann in annotations if isinstance(ann, Points)]
         # KeypointsField expects individual Points objects, not arrays
         # Only supports single keypoint case
@@ -526,7 +531,7 @@ class ForwardEllipseAnnotationConverter(ForwardAnnotationConverter):
             attributes[self.name_prefix + "labels"] = self.ellipse_labels_attribute
         return attributes
 
-    def convert_annotations(self, annotations: list[Annotation], item: DatasetItem) -> dict[str, Any]:
+    def convert_annotations(self, annotations: list[Annotation], item: DatasetItem) -> dict[str, Any]:  # noqa: ARG002
         ellipses: list[list[float]] = []
         labels: list[int | None] = []
 
@@ -603,7 +608,7 @@ class BackwardBboxAnnotationConverter(BackwardAnnotationConverter):
     def get_annotation_type(self) -> AnnotationType:
         return AnnotationType.bbox
 
-    def convert_to_legacy_annotations(self, sample: Sample, categories: CategoriesInfo) -> list[Annotation]:
+    def convert_to_legacy_annotations(self, sample: Sample, categories: CategoriesInfo) -> list[Annotation]:  # noqa: ARG002
         """Convert bboxes and bbox_labels back to legacy Bbox annotations."""
         bboxes = getattr(sample, self.bboxes_attr, None)
         bbox_labels = getattr(sample, self.bbox_labels_attr, None)
@@ -672,7 +677,7 @@ class BackwardRotatedBboxAnnotationConverter(BackwardAnnotationConverter):
     def get_annotation_type(self) -> AnnotationType:
         return AnnotationType.rotated_bbox
 
-    def convert_to_legacy_annotations(self, sample: Sample, categories: CategoriesInfo) -> list[Annotation]:
+    def convert_to_legacy_annotations(self, sample: Sample, categories: CategoriesInfo) -> list[Annotation]:  # noqa: ARG002
         """Convert v2 rotated bbox data to legacy RotatedBbox annotations."""
         rotated_bboxes = getattr(sample, self.rotated_bboxes_attr, None)
         if rotated_bboxes is None or len(rotated_bboxes) == 0:
@@ -758,7 +763,7 @@ class BackwardPolygonAnnotationConverter(BackwardAnnotationConverter):
     def get_annotation_type(self) -> AnnotationType:
         return AnnotationType.polygon
 
-    def convert_to_legacy_annotations(self, sample: Sample, categories: CategoriesInfo) -> list[Annotation]:
+    def convert_to_legacy_annotations(self, sample: Sample, categories: CategoriesInfo) -> list[Annotation]:  # noqa: ARG002
         """Convert polygons and polygon_labels back to legacy Polygon annotations."""
         polygons = getattr(sample, self.polygons_attr)
         polygon_labels = getattr(sample, self.polygon_labels_attr) if self.polygon_labels_attr else None
@@ -923,7 +928,7 @@ class ForwardMaskAnnotationConverter(ForwardAnnotationConverter):
             attributes[self.name_prefix + "labels"] = self.mask_labels_attribute
         return attributes
 
-    def convert_annotations(self, annotations: list[Annotation], item: DatasetItem) -> dict[str, Any]:
+    def convert_annotations(self, annotations: list[Annotation], item: DatasetItem) -> dict[str, Any]:  # noqa: ARG002
         """Convert legacy mask annotations to either semantic or instance segmentation format."""
         # Extract mask annotations
         extracted_masks = [ann for ann in annotations if isinstance(ann, ExtractedMask)]
