+++ src/datumaro/experimental/tiling/tiler_registry.py
@@ -18,9 +18,9 @@ from typing import NamedTuple
 
 import polars as pl
 
-from datumaro.v2.fields import ImageInfoField, TileField, TileInfo
-from datumaro.v2.schema import AttributeInfo, AttributeSpec, Field, Schema, Semantic
-from datumaro.v2.transform import Transform
+from datumaro.experimental.fields import ImageInfoField, TileField, TileInfo
+from datumaro.experimental.schema import AttributeInfo, AttributeSpec, Field, Schema, Semantic
+from datumaro.experimental.transform import Transform
 
 
 @dataclass
