+++ tests/unit/experimental/test_lazy_loading.py
@@ -11,9 +11,9 @@ import polars as pl
 import pytest
 from PIL import Image as PILImage
 
-from datumaro.v2.dataset import Dataset, Sample
-from datumaro.v2.fields import ImageInfo, bbox_field, image_field, image_info_field, image_path_field
-from datumaro.v2.schema import Semantic
+from datumaro.experimental.dataset import Dataset, Sample
+from datumaro.experimental.fields import ImageInfo, bbox_field, image_field, image_info_field, image_path_field
+from datumaro.experimental.schema import Semantic
 
 
 def test_lazy_image_loading_basic():
