+++ tests/unit/experimental/test_sample.py
@@ -8,8 +8,8 @@ import numpy as np
 import polars as pl
 import pytest
 
-from datumaro.v2.dataset import Sample
-from datumaro.v2.fields import (
+from datumaro.experimental.dataset import Sample
+from datumaro.experimental.fields import (
     BBoxField,
     ImageField,
     ImageInfo,
@@ -18,7 +18,7 @@ from datumaro.v2.fields import (
     image_field,
     image_info_field,
 )
-from datumaro.v2.schema import Schema, Semantic
+from datumaro.experimental.schema import Schema, Semantic
 
 
 def test_sample_class_definition():
