+++ tests/unit/experimental/test_tiling.py
@@ -2,9 +2,9 @@ import numpy as np
 import polars as pl
 import pytest
 
-import datumaro.v2.fields.base
-from datumaro.v2.dataset import AttributeInfo
-from datumaro.v2.fields import (
+import datumaro.experimental.fields.base
+from datumaro.experimental.dataset import AttributeInfo
+from datumaro.experimental.fields import (
     TileInfo,
     bbox_field,
     image_field,
@@ -15,8 +15,8 @@ from datumaro.v2.fields import (
     polygon_field,
     tile_field,
 )
-from datumaro.v2.schema import Schema
-from datumaro.v2.tiling.tiler_registry import (
+from datumaro.experimental.schema import Schema
+from datumaro.experimental.tiling.tiler_registry import (
     AttributeSpec,
     TilingConfig,
     _apply_tiling,
