+++ tests/unit/experimental/test_type_registry.py
@@ -10,7 +10,7 @@ from typing import Optional, Union
 import numpy as np
 import pytest
 
-from datumaro.v2.type_registry import from_polars_data, register_from_polars_converter, to_numpy
+from datumaro.experimental.type_registry import from_polars_data, register_from_polars_converter, to_numpy
 
 
 def test_basic_type_conversion():
